Fix an integer overflow bug in the PrimeS code.
It was revealed, embarassingly, by running the PrimeS benchmark in debug
mode.  This code is dirty, and I'll try to clean it up ASAP.

+++ src/common/declust.rs
@@ -96,7 +96,7 @@ pub trait Locator : Send + Sync {
     fn datachunks(&self) -> u64;
 
     /// Number of rows in a single repetition of the layout
-    fn depth(&self) -> i16;
+    fn depth(&self) -> u32;
 
     /// Get an iterator that will iterate through many `ChunkId`s.
     ///
